requirements.txt
langchain
langchain-community
langchain-groq
langchain-google-genai
langchain-chroma
langchain-classic
pypdf
python-dotenv
streamlit